fix: error with circular dependencies
