+++ src/analytics/amplitude.tsx
@@ -1,9 +1,9 @@
-import { getAmpId } from '@/utils/env/client'
+import { getAugmentedAmpId } from '@/providers/config/utils'
 import { createInstance } from '@amplitude/analytics-browser'
 
 export async function createAmplitudeInstance() {
   if (typeof window === 'undefined') return null
-  const ampId = getAmpId()
+  const ampId = getAugmentedAmpId()
   if (!ampId) return null
 
   try {
