+++ src/components/auth/ProfileModal/contents/AccountContent.tsx
@@ -15,7 +15,7 @@ import ProfilePreview from '@/components/ProfilePreview'
 import { SUGGEST_FEATURE_LINK } from '@/constants/links'
 import useFirstVisitNotification from '@/hooks/useFirstVisitNotification'
 import useGetTheme from '@/hooks/useGetTheme'
-import { useConfigContext } from '@/providers/ConfigProvider'
+import { useConfigContext } from '@/providers/config/ConfigProvider'
 import { getLinkedTelegramAccountsQuery } from '@/services/api/notifications/query'
 import { useGetChainDataByNetwork } from '@/services/chainsInfo/query'
 import { getAccountDataQuery } from '@/services/subsocial/evmAddresses'
