+++ src/components/navbar/Navbar/Navbar.tsx
@@ -6,7 +6,7 @@ import { ANN_CHAT_ID } from '@/constants/chat'
 import useIsInIframe from '@/hooks/useIsInIframe'
 import useLoginOption from '@/hooks/useLoginOption'
 import usePrevious from '@/hooks/usePrevious'
-import { useConfigContext } from '@/providers/ConfigProvider'
+import { useConfigContext } from '@/providers/config/ConfigProvider'
 import { getUnreadCountQuery } from '@/services/datahub/posts/query'
 import { useSendEvent } from '@/stores/analytics'
 import { useLoginModal } from '@/stores/login-modal'
