+++ src/modules/chat/ChatPage/StubChatPage.tsx
@@ -2,7 +2,7 @@ import Button from '@/components/Button'
 import { getHubIdFromAlias } from '@/constants/hubs'
 import useToastError from '@/hooks/useToastError'
 import { ApiDiscussionResponse } from '@/pages/api/discussion'
-import { useConfigContext } from '@/providers/ConfigProvider'
+import { useConfigContext } from '@/providers/config/ConfigProvider'
 import { useCreateDiscussion } from '@/services/api/mutation'
 import { getChatPageLink, getUrlQuery } from '@/utils/links'
 import { useRouter } from 'next/router'
